pyproject.toml
[tool.poetry]
name = "hackathon-backend"
version = "0.1.0"
description = ""
authors = ["taivnaaiwnl <taivnaaiwnl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
ruff = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
